Ugh
Replaced Standalone File Browser with yasirkula file browser. SFB only works in .NET 4.0 but the zip utilities for packaging modpacks only works on 2.0. There are 4.0 libraries, but their API's are more difficult to use and not a clean fit.

Look into changing to 4.0 with the updated libraries in the future.